made EvenOddTextReverser a service
